Rename accounts to db
